+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## Unreleased
 
 - Initial documentation.
+- Initial implementation: list view, New, Open (Edit, Title), Complete, Tag,
+  Filter, Settings, Leave.